Restructure `utils.py` to its own package (#915)
* Move utils to its own package

* Move constants to own file

* Update Makefile

* Update forms.py

* Update test_utils.py

* Use absolute imports

* Update 59e9993c169c_change_faces_to_thumbnails.py

* Update .gitignore
