+++ OpenOversight/app/auth/forms.py
@@ -10,7 +10,7 @@ from wtforms.ext.sqlalchemy.fields import QuerySelectField
 from wtforms.validators import DataRequired, Email, EqualTo, Length, Optional, Regexp
 
 from ..models import User
-from ..utils import dept_choices
+from ..utils.db import dept_choices
 
 
 class LoginForm(Form):
