+++ OpenOversight/app/auth/views.py
@@ -11,15 +11,13 @@ from flask import (
 )
 from flask_login import current_user, login_required, login_user, logout_user
 
+from OpenOversight.app.utils.constants import HTTP_METHOD_GET, HTTP_METHOD_POST
+from OpenOversight.app.utils.forms import set_dynamic_default
+from OpenOversight.app.utils.general import validate_redirect_url
+
 from .. import sitemap
 from ..email import send_email
 from ..models import User, db
-from ..utils import (
-    HTTP_METHOD_GET,
-    HTTP_METHOD_POST,
-    set_dynamic_default,
-    validate_redirect_url,
-)
 from . import auth
 from .forms import (
     ChangeDefaultDepartmentForm,
