+++ OpenOversight/app/main/forms.py
@@ -29,9 +29,10 @@ from wtforms.validators import (
     ValidationError,
 )
 
+from OpenOversight.app.utils.db import dept_choices, unit_choices
+
 from ..formfields import TimeField
 from ..models import Officer
-from ..utils import dept_choices, unit_choices
 from ..widgets import BootstrapListWidget, FormFieldWidget
 from .choices import (
     AGE_CHOICES,
