+++ OpenOversight/tests/routes/test_other.py
@@ -4,7 +4,7 @@ from http import HTTPStatus
 import pytest
 from flask import current_app, url_for
 
-from OpenOversight.app.utils import ENCODING_UTF_8
+from OpenOversight.app.utils.constants import ENCODING_UTF_8
 
 from .route_helpers import login_user
